commit1

+++ image_parser.py
@@ -37,7 +37,7 @@ class ImageParser(object):
 
         return size, None
 
-    def size(self, image_path):
+    def size(self, image_path, museum, manifest_file_id):
         # return Image.open(image_path).size
         image_details = defaultdict()
         if image_path in self.url_map:
@@ -45,17 +45,17 @@ class ImageParser(object):
 
         else:
             res = self.getsizes(image_path)
-            try:
-                thumbnail = self.img2base64(image_path)
-            except:
-                thumbnail = "unknown"
-                pass
             image_details["width"] = res[1][0]
             image_details["height"] = res[1][1]
-            image_details["thumbnail"] = thumbnail
-
             self.fob.write(image_path + '\t' + str(image_details["width"]) + '\t' + str(
-                    image_details["height"]) + '\t' + str(thumbnail) + '\n')
+                    image_details["height"]) + '\n')
+
+        try:
+            thumbnail = self.img2base64(image_path, museum, manifest_file_id)
+        except:
+            thumbnail = "unknown"
+            pass
+        image_details["thumbnail"] = thumbnail
         return image_details
 
     def loadFile(self, fileName):
@@ -82,7 +82,11 @@ class ImageParser(object):
     def close(self):
         self.fob.close()
 
-    def img2base64(self, img_link):
+    def img2base64_outdated(self, img_link, museum, manifest_file_id):
+        thumbnailpath = os.path.join("thumbnails", museum, manifest_file_id + ".jpg")
+        print thumbnailpath
+        if os.path.exists(thumbnailpath):
+            return thumbnailpath
         with open("tmp/img_file.jpg", "wb") as f:
             f.write(urllib.urlopen(img_link).read())
         tmp_img = Image.open("tmp/img_file.jpg")
@@ -90,6 +94,22 @@ class ImageParser(object):
         tmp_thumb.save("tmp/thumb_file.jpg")
         with open("tmp/thumb_file.jpg", "rb") as img:
             thumb_string = base64.b64encode(img.read())
-        base64out = "data:image/jpeg;base64," + str(thumb_string)
-        #print base64out
-        return (base64out)
+        # base64out = "data:image/jpeg;base64," + str(thumb_string)
+        base64out = str(thumb_string)
+
+        with open(thumbnailpath, "wb") as f:
+            f.write(base64out)
+        # print base64out
+        # return (base64out)
+        return thumbnailpath
+
+    def img2base64(self, img_link, museum, manifest_file_id):
+        thumbnailpath = os.path.join("thumbnails", museum, manifest_file_id + ".jpg")
+        if os.path.exists(thumbnailpath):
+            return thumbnailpath
+        with open("tmp/img_file.jpg", "wb") as f:
+            f.write(urllib.urlopen(img_link).read())
+        tmp_img = Image.open("tmp/img_file.jpg")
+        tmp_thumb = tmp_img.resize((250, 250), Image.ANTIALIAS)
+        tmp_thumb.save(thumbnailpath)
+        return thumbnailpath
